Add optional SVG Normalizer
With this change, we can make all svg elements to path, and handle it. It creates a flat SVG, so no grouping or something else will take affect, its the same way currently the parser will also handle the file preperation.

A new Option is added to the config ("Use SVG Normalizer?") it can enabled and disabled.